% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ask_user}
\alias{ask_user}
\title{Title}
\usage{
ask_user(question = "Proceed?", default = FALSE, ask = interactive())
}
\arguments{
\item{question}{QUESTION to ask}

\item{default}{default reply, true or false}

\item{ask}{will be asked only if interactive}
}
\value{
boolean true or false
}
\description{
Title
}
\examples{
# none
}
