% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_set_lasR_cores.R
\name{auto_set_lasR_cores}
\alias{auto_set_lasR_cores}
\title{auto_set_lasR_cores}
\usage{
auto_set_lasR_cores(ctg, safety = 0.7, inflate = 3, verbose = FALSE)
}
\arguments{
\item{ctg}{\itemize{
\item a LAScatalog
}}

\item{safety}{\itemize{
\item percent of RAM you allow to be used (0.8 = 80\%)
}}

\item{inflate}{\itemize{
\item multiplier for memory inflation (metadata, temp objects, overhead)
}}

\item{verbose}{bolean default = TRUE - print lots of messages.}
}
\value{
the number of safe cores
}
\description{
auto_set_lasR_cores
}
\examples{
# NONE
}
