% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_clear}
\alias{cache_clear}
\title{cache_clear}
\usage{
cache_clear(silent = F)
}
\arguments{
\item{silent}{bolean: default is FALSE so that the user is asked confirmation.
If TRUE, the cache file will be cleared directly without any confirmation request.}
}
\value{
environment cache of package
}
\description{
Removes all information from cache saved in envir .lasRpipeline_cache
}
\examples{
# cache_clear()
}
