% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{check_dir_structure}
\alias{check_dir_structure}
\title{check_dir_structure}
\usage{
check_dir_structure(ifiles = NULL, odir = NULL)
}
\arguments{
\item{ifiles}{input files or directory with LAS/LAZ files}

\item{odir}{output directory}
}
\value{
TRUE if all OK
}
\description{
check_dir_structure
}
\examples{
#
}
