% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{check_vrt_validity}
\alias{check_vrt_validity}
\title{check_vrt_validity}
\usage{
check_vrt_validity(filevrt)
}
\arguments{
\item{filevrt}{input files or directory with LAS/LAZ files}
}
\value{
raster pointer if all OK, FALSE otherwise
}
\description{
check_vrt_validity
}
\examples{
#
}
