% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache}
\alias{get_cache}
\title{get_cache}
\usage{
get_cache(key)
}
\arguments{
\item{key}{character with key of cache}
}
\value{
value of  cached  key
}
\description{
Return cache using key
}
\examples{
get_cache("laspattern")
}
