% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_env_ls}
\alias{get_cache_env_ls}
\title{get_cache_env_ls}
\usage{
get_cache_env_ls()
}
\value{
environment cache of package
}
\description{
Return some cached info related to point clouds so that they don't have to
be recalculated.
}
\examples{
get_cache_env_ls()
}
