% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasHAGinfo.R
\name{hasHAGinfo}
\alias{hasHAGinfo}
\title{hasGroundPoints}
\usage{
hasHAGinfo(f, verbose = T)
}
\arguments{
\item{f}{file name or character vector with file names to process}

\item{verbose}{if TRUE will provide verbose messages, if FALSE then it will be silent.}
}
\value{
FALSE if no classified ground points are present or
a numeric with the ratio of points classified as ground in the file
}
\description{
Will return TRUE if a sampled las file has Height Above Ground information,
which means it has been normalized already.
}
\examples{
testFile <- system.file("extdata", "BL5_UTM_32_ort_0021.laz", package = "lasRpipeline")
hasGroundPoints(testFile)
}
