% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{list_files_still_to_process}
\alias{list_files_still_to_process}
\title{Title}
\usage{
list_files_still_to_process(f, odir, inverse = F)
}
\arguments{
\item{f}{original files, can be a lidR catalogue or a folder with LAS/LAZ or a vector of full paths to LAS/LAZ files}

\item{odir}{output directory where supposedly there are the processed files, e.g. for normalized files it could be "output/norm"}

\item{inverse}{return files that have been already processed instead of files to process}
}
\value{
a vector with full paths to files that still have to be processed, or that have been processed (if inverse=T)
}
\description{
Title
}
\examples{
# none
}
