% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_log.R
\name{message_log}
\alias{message_log}
\title{Title}
\usage{
message_log(..., isWarning = F)
}
\arguments{
\item{...}{characters to add to the message}

\item{isWarning}{if TRUE then a warning message is created and both a
warning.log and message.log file is updated.}
}
\value{
files names with pid and warning.log and message.log.
}
\description{
Generate a diagnostic message from its arguments just like
\href{library/base/help/warning}{message} except that it logs it to a file in
the directory "logs" that will have the PID number of the process
and extension warning.log,
e.g. NNNNNNN_warning.log
}
\examples{
# message_log("my nice ", "message in file with PID=", Sys.getpid())
}
