% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(f, odir = "outdir/norm", force = FALSE, verbose = TRUE)
}
\arguments{
\item{f}{file or list of files to process}

\item{odir}{output directory, default = outdir/norm}

\item{force}{force rewrite of files in output}

\item{verbose}{verbose output messages in logs}
}
\value{
list of normalized files with full path, that can be used for
successive processing.
}
\description{
Adds height a above ground of points, with heuristics on existing temp files
in the project
}
\examples{
# none
}
