% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShade.R
\name{plotShade}
\alias{plotShade}
\title{Title}
\usage{
plotShade(
  dtm,
  dtm_cols = grDevices::terrain.colors(100),
  main = "DEM with Hillshade"
)
}
\arguments{
\item{dtm}{terra raster with DTM}

\item{dtm_cols}{colors - default is grDevices::terrain.colors(100)}

\item{main}{the title - default is ""}
}
\description{
Title
}
\examples{
# none
}
