% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(
  ifiles = NULL,
  odir = NULL,
  gridfile = NULL,
  create = list(createDTM = TRUE, createCHM = TRUE),
  maxh = NULL,
  forceRes = NULL,
  verbose = TRUE,
  progress = TRUE,
  ncores = NULL,
  buffer = NULL,
  chunk = NULL
)
}
\arguments{
\item{ifiles}{path with input las/laz files or specific list of las/laz files' path
if not provided it will search the current working directory.}

\item{odir}{path to output directory - if not provided it will create an "odir"
directory with name "odir" in the current working directory.}

\item{gridfile}{path to raster that is the template of your output - MUST
overlap the lidar area of course}

\item{create}{list with following
Must include:
\describe{
\item{createDTM}{Bolean: default = TRUE}
\item{createCHM}{Bolean: default = TRUE}
}}

\item{maxh}{integer, if set it will set a maximum height above the ground
to consider it as vegetation.}

\item{forceRes}{default = NULL - if number is given, it will force resolution
of the output intermediate rasters (DTM CHM) to this value, otherwise it will
use the resolution of raster in \code{gridfile}}

\item{verbose}{Boolean default = TRUE - print lots of messages.}

\item{progress}{Boolean. Displays a progress bar.}

\item{ncores}{An object returned by one of \code{sequential()},
\code{concurrent_points()}, \code{concurrent_files()}, or \code{nested()}.
If \code{NULL} the default it will calculate how many
concurrent LAS files to process by trying to check the RAM and number of cores
and divide it by the maximum size of LAS files. If a simple integer is provided
it corresponds to \code{concurrent_files(ncores)}.}

\item{buffer}{Numeric. Each file is read with a buffer. The default is
\code{NULL}, which does not mean that the file will not be buffered.
It means that the internal routine determines whether a buffer is needed
and will pick the greatest value between its internal suggestion and this
user-supplied value.}

\item{chunk}{Numeric. By default, files are processed one by one
(\code{chunk = NULL} or \code{chunk = 0}). It is possible to process in
arbitrary-sized chunks. This is useful for processing collections with
large files or for handling massive \code{copc} files.}
}
\value{
a terra raster object also written to a file tif with time and date,
written to the odir e.g.  "gridOut20250811_094743.tif"
The TIF file will have seven layers, with:
\itemize{
\item terrain height/slope/aspect
\item canopy cover (0-100 percent)
\item canopy height (95th percentile)
\item crown base height (m)
\item canopy bulk density (kg/m3)
}
}
\description{
This will launch optimized processes using 20 plots per chunk and parallelization

Works only in linux for now!

Plots are the centers of the raster cell where we want to calculate topography
and canopy information.
}
\examples{
 #

 create = list(
 createDTM = TRUE,
 createCHM = TRUE
 )
 maxh = NULL
 forceRes = 1
 verbose=TRUE
 buffer=NULL
 ncores=NULL
 chunk=NULL
 progress=TRUE

}
