% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuelMetrics.R
\name{slope_aspect}
\alias{slope_aspect}
\title{slope_aspect}
\usage{
slope_aspect(x, y, z)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}
}
\value{
a list with slope and aspect in degrees, both with
lme and wi
}
\description{
Return interpolated slope and aspect from linear model and
robust linear model
}
\examples{
# none
}
