% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(
  f,
  odir,
  gridfile = "../wildfire/input/AT-IT_ScottBurganFuelMapClassV2.tif"
)
}
\arguments{
\item{f}{path with input las/laz files or specific list of las/laz files}

\item{odir}{path to output directory}

\item{gridfile}{path to raster that is the template of your output - MUST overlap the lidar area of course}
}
\value{
NA
}
\description{
process
}
\examples{
 # use this to process example
}
